Update Covid_Tracker_Final.cpp
New deaths over a period of time

I fixed data sorting that happens after top cases/deaths functions and noticed that it doesn't make sense to search on period of dates.

Row with maximum cases/deaths for each state will be on the last day of the period anyway. So just search by exact date will be enough

Also it now displays actual numbers, not just the number of times each state has occurred in the CSV